# openapi.yaml
openapi: 3.0.3
info:
  title: Cookpad-like Recipe Sharing API - Microservices Architecture
  description: |
    API cho ứng dụng chia sẻ công thức nấu ăn, tương tự Cookpad.
    
    **Kiến trúc Microservices:**
    - API Gateway: http://localhost:8888/api/v1/...
    
    Tính năng chính:
    - Quản lý người dùng và xác thực
    - Tạo, chỉnh sửa, xóa công thức nấu ăn
    - Tìm kiếm và lọc công thức
    - Đánh giá và bình luận
    - Yêu thích và lưu công thức
    - Theo dõi người dùng khác
    - Quản lý danh mục và thẻ tag
    - Phân tích hồ sơ bệnh án và gợi ý món ăn AI
    
    Vào endpoint sau để xem Swagger UI: http://localhost:5000/api/docs/ (đăng nhập và truyền token vào Authorize)
  version: 1.0.0
servers:
  - url: http://localhost:8888/api/v1
    description: API Gateway - Local Development
...
paths:
...
  /tags:
    get:
      tags:
        - Tag Service
      summary: Lấy danh sách tags
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Tìm kiếm tag theo tên
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Danh sách tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'

    post:
      tags:
        - Tag Service
      summary: Tạo tag mới
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: healthy
      responses:
        '200':
          description: Tạo tag thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token để xác thực. Format "Bearer {token}"
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Số lượng items mỗi trang
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: string
          example: tag_xyz789
        name:
          type: string
          example: healthy
        recipesCount:
          type: integer
          example: 456
        createdAt:
          type: string
          format: date-time
...